requirements.txt
faiss-cpu>=1.7.0
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0
tqdm>=4.60.0